move chunk graph building into separate file
